修改点赞接口、增加获取点赞(浏览)状态、通用发帖和发送图片接口整合，失物发帖和图片接口整合

+++ src/main/java/net/gupt/community/controller/StudentController.java
@@ -79,7 +79,7 @@ public class StudentController {
      * @return 结果集输出信息
      */
     @RequestMapping(value = "/getMyArticles", method = RequestMethod.GET)
-    public Result getMyArticles(@RequestParam(value = "postType") Integer postType,
+    public Result getMyArticles(@RequestParam(value = "postType") Byte postType,
                                 @RequestParam(value = "pageNum") Integer pageNum,
                                 @RequestParam(value = "pageSize") Integer pageSize,
                                 @RequestParam(value = "id", required = false) Integer id,
